backend/requirements.txt
FastAPI
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
bcrypt
httpx
pydantic